[master]活动详情查询接口出参组装

+++ nas-common/src/main/java/com/minyan/nascommon/po/ActivityEventPO.java
@@ -27,7 +27,7 @@ public class ActivityEventPO implements Serializable {
     /**
      * 模块id
      */
-    private Long moduleId;
+    private Integer moduleId;
 
     /**
      * 事件名称
